Use appropriate size

+++ vsc7448-pac/src/lib.rs
@@ -278,7 +278,7 @@ impl Vsc7448 {
 /// The Automatic Frame Injector (AFI) provides mechanisms for periodic injection of PDUs. E.g. injection of - OAM PDUs for continuity check, loss and delay measurement - OAM PDUs for high service activation test, ref. ITU Y.1564. - IEEE 1588 PDUs
 pub struct AFI(u32);
 impl AFI {
-    pub const BASE: usize = 0x71a80000;
+    pub const BASE: u32 = 0x71a80000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -329,7 +329,7 @@ impl AFI {
 /// Controls mask handling etc.
 pub struct ANA_AC(u32);
 impl ANA_AC {
-    pub const BASE: usize = 0x71f00000;
+    pub const BASE: u32 = 0x71f00000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -467,7 +467,7 @@ impl ANA_AC {
 /// Access Control List sub block of the Analyzer
 pub struct ANA_ACL(u32);
 impl ANA_ACL {
-    pub const BASE: usize = 0x71430000;
+    pub const BASE: u32 = 0x71430000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -500,7 +500,7 @@ impl ANA_ACL {
 /// This block os part of the Vitesse OAM MEP Processor (VOP). It is directly controlled by the VOP and performs all frame modification related OAM functions. This block is included in both the REW and the ANA_AC. The HW support for implementing an OAM MEP is implemented in a sub block, known as: Vitesse OAM Endpoint (VOE). The VOP includes the following: * 256 Service / Path VOEs * 11 Port VOEs
 pub struct ANA_AC_OAM_MOD(u32);
 impl ANA_AC_OAM_MOD {
-    pub const BASE: usize = 0x71b80000;
+    pub const BASE: u32 = 0x71b80000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -529,7 +529,7 @@ impl ANA_AC_OAM_MOD {
 /// Controls the various policers.
 pub struct ANA_AC_POL(u32);
 impl ANA_AC_POL {
-    pub const BASE: usize = 0x71980000;
+    pub const BASE: u32 = 0x71980000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -584,7 +584,7 @@ impl ANA_AC_POL {
 /// Classifier sub block of the Analyzer
 pub struct ANA_CL(u32);
 impl ANA_CL {
-    pub const BASE: usize = 0x71d00000;
+    pub const BASE: u32 = 0x71d00000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -641,7 +641,7 @@ impl ANA_CL {
 /// Layer 2 sub block of the Analyzer
 pub struct ANA_L2(u32);
 impl ANA_L2 {
-    pub const BASE: usize = 0x71e00000;
+    pub const BASE: u32 = 0x71e00000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -679,7 +679,7 @@ impl ANA_L2 {
 /// ANA_L3 provides the following functionality: - VLAN handling - MSTP handling - IP routing
 pub struct ANA_L3(u32);
 impl ANA_L3 {
-    pub const BASE: usize = 0x71800000;
+    pub const BASE: u32 = 0x71800000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -740,7 +740,7 @@ impl ANA_L3 {
 /// Assembler
 pub struct ASM(u32);
 impl ASM {
-    pub const BASE: usize = 0x71410000;
+    pub const BASE: u32 = 0x71410000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -792,8 +792,8 @@ impl ASM {
 /// \see vtss_target_DEV10G_e VAUI @ 12Gbps 2 lanes, XAUI @ 10 Gbps, XAUI @ 12 Gbps, VAUI @ 10Gbps 2 lanes and SGMII @ 10/100/1000/2500 Mbps
 pub struct DEV10G(u32);
 impl DEV10G {
-    pub const BASE: usize = 0x71390000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x71390000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -847,8 +847,8 @@ impl DEV10G {
 /// \see vtss_target_DEV1G_e The device operates as an interface between the TAXI bus and SGMII macro supporting data rates of 10M, 100M, and 1000Mbps. The device includes PCS, MAC and rate adaption logic.
 pub struct DEV1G(u32);
 impl DEV1G {
-    pub const BASE: usize = 0x71040000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x71040000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -886,8 +886,8 @@ impl DEV1G {
 /// \see vtss_target_DEV1G_e The device operates as an interface between the TAXI bus and SGMII macro supporting data rates of 10M, 100M, and 1000Mbps. The device includes PCS, MAC and rate adaption logic.
 pub struct DEV2G5(u32);
 impl DEV2G5 {
-    pub const BASE: usize = 0x711c0000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x711c0000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -925,7 +925,7 @@ impl DEV2G5 {
 /// General Configuration Block
 pub struct DEVCPU_GCB(u32);
 impl DEVCPU_GCB {
-    pub const BASE: usize = 0x71010000;
+    pub const BASE: u32 = 0x71010000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -985,7 +985,7 @@ impl DEVCPU_GCB {
 /// CPU device origin
 pub struct DEVCPU_ORG(u32);
 impl DEVCPU_ORG {
-    pub const BASE: usize = 0x71000000;
+    pub const BASE: u32 = 0x71000000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -999,7 +999,7 @@ impl DEVCPU_ORG {
 /// DEVCPU Precision Timing Protocol Originator
 pub struct DEVCPU_PTP(u32);
 impl DEVCPU_PTP {
-    pub const BASE: usize = 0x717f0000;
+    pub const BASE: u32 = 0x717f0000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1023,7 +1023,7 @@ impl DEVCPU_PTP {
 /// CPU Device Queue System
 pub struct DEVCPU_QS(u32);
 impl DEVCPU_QS {
-    pub const BASE: usize = 0x71020000;
+    pub const BASE: u32 = 0x71020000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1041,7 +1041,7 @@ impl DEVCPU_QS {
 /// Disassembler
 pub struct DSM(u32);
 impl DSM {
-    pub const BASE: usize = 0x71450000;
+    pub const BASE: u32 = 0x71450000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1075,7 +1075,7 @@ impl DSM {
 /// Hierarchical Scheduler Configuration
 pub struct HSCH(u32);
 impl HSCH {
-    pub const BASE: usize = 0x71880000;
+    pub const BASE: u32 = 0x71880000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1144,7 +1144,7 @@ impl HSCH {
 /// Register Collection for Control of Macros (SERDES1G, SERDES6G, LCPLL)
 pub struct HSIO(u32);
 impl HSIO {
-    pub const BASE: usize = 0x71460000;
+    pub const BASE: u32 = 0x71460000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1230,7 +1230,7 @@ impl HSIO {
 /// VCore Configuration
 pub struct ICPU_CFG(u32);
 impl ICPU_CFG {
-    pub const BASE: usize = 0x70000000;
+    pub const BASE: u32 = 0x70000000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1288,7 +1288,7 @@ impl ICPU_CFG {
 /// Learn block
 pub struct LRN(u32);
 impl LRN {
-    pub const BASE: usize = 0x71420000;
+    pub const BASE: u32 = 0x71420000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1302,7 +1302,7 @@ impl LRN {
 /// PCIe Endpoint Configuration Space
 pub struct PCIE(u32);
 impl PCIE {
-    pub const BASE: usize = 0x70111000;
+    pub const BASE: u32 = 0x70111000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1336,8 +1336,8 @@ impl PCIE {
 /// \see vtss_target_PCS_10GBASE_R_e PCS configuration and status registers and counters
 pub struct PCS10G_BR(u32);
 impl PCS10G_BR {
-    pub const BASE: usize = 0x713d0000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x713d0000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1387,7 +1387,7 @@ impl PCS10G_BR {
 /// Queue System Configuration
 pub struct QFWD(u32);
 impl QFWD {
-    pub const BASE: usize = 0x717d0000;
+    pub const BASE: u32 = 0x717d0000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1401,7 +1401,7 @@ impl QFWD {
 /// Queue System Configuration
 pub struct QRES(u32);
 impl QRES {
-    pub const BASE: usize = 0x71900000;
+    pub const BASE: u32 = 0x71900000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1425,7 +1425,7 @@ impl QRES {
 /// Queue System Configuration
 pub struct QSYS(u32);
 impl QSYS {
-    pub const BASE: usize = 0x717e0000;
+    pub const BASE: u32 = 0x717e0000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1467,7 +1467,7 @@ impl QSYS {
 /// Rewriter
 pub struct REW(u32);
 impl REW {
-    pub const BASE: usize = 0x71b00000;
+    pub const BASE: u32 = 0x71b00000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1551,7 +1551,7 @@ impl REW {
 /// Shared Bus Arbiter
 pub struct SBA(u32);
 impl SBA {
-    pub const BASE: usize = 0x70110000;
+    pub const BASE: u32 = 0x70110000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1565,7 +1565,7 @@ impl SBA {
 /// SI Master Controller
 pub struct SIMC(u32);
 impl SIMC {
-    pub const BASE: usize = 0x70101000;
+    pub const BASE: u32 = 0x70101000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1579,7 +1579,7 @@ impl SIMC {
 /// \see vtss_target_TWI_e Two-Wire Interface Controller
 pub struct TWI(u32);
 impl TWI {
-    pub const BASE: usize = 0x70100400;
+    pub const BASE: u32 = 0x70100400;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1593,7 +1593,7 @@ impl TWI {
 /// \see vtss_target_TWI_e Two-Wire Interface Controller
 pub struct TWI2(u32);
 impl TWI2 {
-    pub const BASE: usize = 0x70100c00;
+    pub const BASE: u32 = 0x70100c00;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1607,7 +1607,7 @@ impl TWI2 {
 /// \see vtss_target_UART_e UART Controller
 pub struct UART(u32);
 impl UART {
-    pub const BASE: usize = 0x70100000;
+    pub const BASE: u32 = 0x70100000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1621,7 +1621,7 @@ impl UART {
 /// \see vtss_target_UART_e UART Controller
 pub struct UART2(u32);
 impl UART2 {
-    pub const BASE: usize = 0x70100800;
+    pub const BASE: u32 = 0x70100800;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1635,7 +1635,7 @@ impl UART2 {
 /// \see vtss_target_VAUI_CHANNEL_e Vaui channel configuration and status register set
 pub struct VAUI0(u32);
 impl VAUI0 {
-    pub const BASE: usize = 0x717b0000;
+    pub const BASE: u32 = 0x717b0000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1659,7 +1659,7 @@ impl VAUI0 {
 /// \see vtss_target_VAUI_CHANNEL_e Vaui channel configuration and status register set
 pub struct VAUI1(u32);
 impl VAUI1 {
-    pub const BASE: usize = 0x717c0000;
+    pub const BASE: u32 = 0x717c0000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1683,7 +1683,7 @@ impl VAUI1 {
 /// \see vtss_target_VCAP_CORE_e Vitesse Content Aware Processor
 pub struct VCAP_ES0(u32);
 impl VCAP_ES0 {
-    pub const BASE: usize = 0x71470000;
+    pub const BASE: u32 = 0x71470000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1717,7 +1717,7 @@ impl VCAP_ES0 {
 /// \see vtss_target_VCAP_CORE_e Vitesse Content Aware Processor
 pub struct VCAP_SUPER(u32);
 impl VCAP_SUPER {
-    pub const BASE: usize = 0x71440000;
+    pub const BASE: u32 = 0x71440000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1751,7 +1751,7 @@ impl VCAP_SUPER {
 /// The Vitesse OAM MEP Processor (VOP) implements the HW support for implementing OAM MEPs. The HW support for implementing an OAM MEP is implemented in a sub block, referred to as: Vitesse OAM Endpoint (VOE). The VOEs can be configured for either Ethernet PDU processing or MPLS-TP PDU processing. This CSR target us used for configuring VOEs configured for Ethernet processing. The VOP includes the following: * 256 Service / Path VOEs * 11 Port VOEs (10 front ports + NPI) VOEs are configured for either Ethernet or MPLS-TP support using the following register: * VOP:VOE_CONF_REG:VOE_MISC_CONFIG.MPLS_OAM_ENA Use CSR target VOP to configure VOEs configured for Ethernet. To configure VOEs configured for MPLS-TP, use CSR target: VOP_MPLS. For VOEs configured for MPLS-TP, the following register groups are invalid: * VOP:VOE_CONF:* * VOP:VOE_STAT:*
 pub struct VOP(u32);
 impl VOP {
-    pub const BASE: usize = 0x71c00000;
+    pub const BASE: u32 = 0x71c00000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1828,7 +1828,7 @@ impl VOP {
 /// The Vitesse OAM MEP Processor (VOP) implements the HW support for implementing OAM MEPs. The HW support for implementing an OAM MEP is implemented in a sub block, referred to as: Vitesse OAM Endpoint (VOE). The VOEs can be configured for either Ethernet PDU processing or MPLS-TP PDU processing. This CSR target us used for configuring VOEs configured for MPLS-TP processing. The VOP includes the following: * 256 Service / Path VOEs * 11 Port VOEs (11 front ports) VOEs are configured for either Ethernet or MPLS-TP support using the following register: * VOP:VOE_CONF_REG:VOE_MISC_CONFIG.MPLS_OAM_ENA Use CSR target VOP_MPLS to configure VOEs configured for MPLS_TP. To configure VOEs configured for Ethernet use CSR target: VOP. For VOEs configured for Ethernet, the following register groups are invalid: * VOP:CONF_MPLS:* * VOP:STAT_MPLS:* Note that the VOP_MPLS CSR target shares the same physical RAM as VOP CSR target. The RAM is initialized to the VOP CSR target values (Ethernet VOE). Hence the initial values listed in the VOP_MPLS are not valid. Prior to using a VOE configured for MPLS, all REGISTERS MUST BE INITIALIZED.
 pub struct VOP_MPLS(u32);
 impl VOP_MPLS {
-    pub const BASE: usize = 0x71a00000;
+    pub const BASE: u32 = 0x71a00000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1848,8 +1848,8 @@ impl VOP_MPLS {
 /// \see vtss_target_SD10G65_e Configuration and status for high speed macro.
 pub struct XGANA(u32);
 impl XGANA {
-    pub const BASE: usize = 0x71480000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x71480000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1887,8 +1887,8 @@ impl XGANA {
 /// \see vtss_target_SD10G65_DIG_e Configuration and status of digital peripherals for high speed macro.
 pub struct XGDIG(u32);
 impl XGDIG {
-    pub const BASE: usize = 0x714c0000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x714c0000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -1918,8 +1918,8 @@ impl XGDIG {
 /// \see vtss_target_KR_DEV1_e
 pub struct XGKR0(u32);
 impl XGKR0 {
-    pub const BASE: usize = 0x71540000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x71540000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -2081,8 +2081,8 @@ impl XGKR0 {
 /// \see vtss_target_KR_DEV7_e
 pub struct XGKR1(u32);
 impl XGKR1 {
-    pub const BASE: usize = 0x71580000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x71580000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -2184,8 +2184,8 @@ impl XGKR1 {
 /// \see vtss_target_XFI_SHELL_e XFI shell
 pub struct XGXFI(u32);
 impl XGXFI {
-    pub const BASE: usize = 0x71500000;
-    pub const SIZE: usize = 0x10000;
+    pub const BASE: u32 = 0x71500000;
+    pub const SIZE: u32 = 0x10000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
@@ -2199,7 +2199,7 @@ impl XGXFI {
 /// Queue Transfer Configuration
 pub struct XQS(u32);
 impl XQS {
-    pub const BASE: usize = 0x717a0000;
+    pub const BASE: u32 = 0x717a0000;
     pub unsafe fn new_unchecked(a: u32) -> Self {
         Self(a)
     }
